+++ src/module/helpers/settings.ts
@@ -24,10 +24,21 @@ export class IronswornSettings {
       type: Boolean,
       default: true,
     })
+
+    game.settings.register('foundry-ironsworn', 'data-version', {
+      scope: 'world',
+      config: false,
+      type: Number,
+      default: 1,
+    })
   }
 
   static get theme(): string {
     return game.settings.get('foundry-ironsworn', 'theme') as string
+  }
+
+  static get dataVersion(): number {
+    return game.settings.get('foundry-ironsworn', 'data-version') as number
   }
 
   static async maybeSetGlobalSupply(value: number) {